A routine to find all matches given function like 'find-tag-noselect.
The idea is to be able to reuse existing code for matching tags
instead of doing all the hard work on our own. For instance, some
major modes has their own idea of what matched tag
means (e.g. erlang-mode). With this approach the code of such a mode
can be easily reused. On the other hand, there's also a serious
drawback. Namely, instead of returning some kind of meta-information
about tag match, functions like 'find-tag-noselect return a buffer
with a mark set to the match. Thus in order to build a selection
buffer, we'll have to visit all the matching files. This might be less
than desirable if there're lots of matches. Anyway, will see how it
works in practice.